# #115 - Avoid array insertion in iterative refinement