--- a/src/Approximations/iterative_refinement.jl
+++ b/src/Approximations/iterative_refinement.jl
@@ -5,13 +5,13 @@ Type that represents a local approximation in 2D.
 
 ### Fields
 
-- `p1`         -- first inner point
-- `d1`         -- first direction
-- `p2`         -- second inner point
-- `d2`         -- second direction
-- `q`          -- intersection of the lines l1 ⟂ d1 at p1 and l2 ⟂ d2 at p2
-- `refinable`  -- states if this approximation is refinable
-- `err`        -- error upper bound
+- `p1`        -- first inner point
+- `d1`        -- first direction
+- `p2`        -- second inner point
+- `d2`        -- second direction
+- `q`         -- intersection of the lines l1 ⟂ d1 at p1 and l2 ⟂ d2 at p2
+- `refinable` -- states if this approximation is refinable
+- `err`       -- error upper bound
 
 ### Notes
 
@@ -34,14 +34,14 @@ Type that represents the polygonal approximation of a convex set.
 
 ### Fields
 
-- `S`                -- convex set
-- `approx_list_open` -- vector of local approximations that are not finished yet
-- `approx_list_done` -- vector of linear constraints that are already finished
+- `S`           -- convex set
+- `approx_list` -- vector of local approximations that are not finished yet
+- `constraints` -- vector of linear constraints that are already finished
 """
 struct PolygonalOverapproximation{N<:Real}
     S::LazySet{N}
-    approx_list_open::Vector{LocalApproximation{N}}
-    approx_list_done::Vector{LinearConstraint{N}}
+    approx_list::Vector{LocalApproximation{N}}
+    constraints::Vector{LinearConstraint{N}}
 
     PolygonalOverapproximation(S::LazySet{N}) where N<:Real = new{N}(
         S, Vector{LocalApproximation{N}}(), Vector{LinearConstraint{N}}())
@@ -53,11 +53,11 @@ end
 
 ### Input
 
-- `S`          -- convex set
-- `p1`         -- first inner point
-- `d1`         -- first direction
-- `p2`         -- second inner point
-- `d2`         -- second direction
+- `S`  -- convex set
+- `p1` -- first inner point
+- `d1` -- first direction
+- `p2` -- second inner point
+- `d2` -- second direction
 
 ### Output
 
@@ -72,8 +72,8 @@ function new_approx(S::LazySet, p1::Vector{N}, d1::Vector{N}, p2::Vector{N},
         ndir = normalize([p2[2]-p1[2], p1[1]-p2[1]])
         q = element(intersection(Line(d1, dot(d1, p1)), Line(d2, dot(d2, p2))))
         approx_error = min(norm(q - σ(ndir, S)), dot(ndir, q - p1))
-        refinable = (approx_error > TOL(N)) &&
-                     !(norm(p1-q, 2) <= TOL(N) || norm(q-p2, 2) <= TOL(N))
+        refinable = (approx_error > TOL(N)) && (norm(p1-q, 2) > TOL(N)) &&
+                    (norm(q-p2, 2) > TOL(N))
         ap = LocalApproximation{N}(p1, d1, p2, d2, q, refinable, approx_error)
     end
     return ap
@@ -85,11 +85,11 @@ end
 
 ### Input
 
-- `Ω`          -- polygonal overapproximation of a convex set
-- `p1`         -- first inner point
-- `d1`         -- first direction
-- `p2`         -- second inner point
-- `d2`         -- second direction
+- `Ω`  -- polygonal overapproximation of a convex set
+- `p1` -- first inner point
+- `d1` -- first direction
+- `p2` -- second inner point
+- `d2` -- second direction
 
 ### Output
 
@@ -101,7 +101,7 @@ function addapproximation!(Ω::PolygonalOverapproximation,
                            d2::Vector{N})::LocalApproximation{N} where N<:Real
 
     approx = new_approx(Ω.S, p1, d1, p2, d2)
-    push!(Ω.approx_list_open, approx)
+    push!(Ω.approx_list, approx)
     return approx
 end
 
@@ -110,13 +110,13 @@ end
            R::LocalApproximation
           )::Tuple{LocalApproximation, LocalApproximation}
 
-Refine a given local approximation of the polygonal approximation of a convex set,
-by splitting along the normal direction to the approximation.
+Refine a given local approximation of the polygonal approximation of a convex
+set by splitting along the normal direction to the approximation.
 
 ### Input
 
-- `Ω`   -- polygonal overapproximation of a convex set
-- `i`   -- integer index for the local approximation to be refined
+- `Ω` -- polygonal overapproximation of a convex set
+- `i` -- integer index for the local approximation to be refined
 
 ### Output
 
@@ -135,20 +135,21 @@ function refine(Ω::PolygonalOverapproximation,
 end
 
 """
-    tohrep(Ω::PolygonalOverapproximation{N})::AbstractHPolygon where N<:Real
+    tohrep(Ω::PolygonalOverapproximation{N})::AbstractHPolygon{N} where N<:Real
 
 Convert a polygonal overapproximation into a concrete polygon.
 
 ### Input
 
-- `Ω`   -- polygonal overapproximation of a convex set
+- `Ω` -- polygonal overapproximation of a convex set
 
 ### Output
 
 A polygon in constraint representation.
 """
-function tohrep(Ω::PolygonalOverapproximation{N})::AbstractHPolygon where N<:Real
-    return HPolygon{N}(Ω.approx_list_done)
+function tohrep(Ω::PolygonalOverapproximation{N}
+               )::AbstractHPolygon{N} where N<:Real
+    return HPolygon{N}(Ω.constraints)
 end
 
 """
@@ -179,39 +180,42 @@ function approximate(S::LazySet{N},
 
     Ω = PolygonalOverapproximation(S)
 
+    # add constraints in reverse (i.e., clockwise) order
     addapproximation!(Ω, ps, DIR_SOUTH(N), pe, DIR_EAST(N))
     addapproximation!(Ω, pw, DIR_WEST(N), ps, DIR_SOUTH(N))
     addapproximation!(Ω, pn, DIR_NORTH(N), pw, DIR_WEST(N))
     addapproximation!(Ω, pe, DIR_EAST(N), pn, DIR_NORTH(N))
 
-    open = Ω.approx_list_open
-    done = Ω.approx_list_done
-    while !isempty(open)
-        approx = pop!(open)
+    approx_list = Ω.approx_list
+    while !isempty(approx_list)
+        approx = pop!(approx_list)
 
         if !approx.refinable || approx.err <= ε
             # if the approximation is not refinable => continue
-            push!(done, LinearConstraint(approx.d1, dot(approx.d1, approx.p1)))
+            push!(Ω.constraints,
+                  LinearConstraint(approx.d1, dot(approx.d1, approx.p1)))
             continue
         end
 
+        # refine
         (la1, la2) = refine(Ω, approx)
 
-        if isempty(open)
+        if isempty(approx_list)
             redundant = false
         else
-            next = open[end]
+            # check if the next local approximation became redundant
+            next = approx_list[end]
             redundant = (norm(la2.p1-next.p1) <= TOL(N)) &&
                 (norm(la2.q-next.q) <= TOL(N))
         end
         if redundant
             # replace redundant old constraint
-            open[end] = la2
+            approx_list[end] = la2
         else
             # add new constraint
-            push!(open, la2)
+            push!(approx_list, la2)
         end
-        push!(open, la1)
+        push!(approx_list, la1)
     end
     return Ω
 end