Closes #115.