@schillic 2018-10-26
The new implementation is only more efficient for very large numbers of constraints.

```julia
using LazySets, LazySets.Approximations, BenchmarkTools
B = Ball2(zeros(2), 1.);


# master branch:

julia> @btime overapproximate(B, 1e-2);
  104.092 μs (2573 allocations: 132.77 KiB)

julia> @btime overapproximate(B, 1e-5);
  3.959 ms (88850 allocations: 4.51 MiB)

julia> @btime overapproximate(B, 1e-8);
  225.120 ms (2977460 allocations: 147.42 MiB)


# this branch:

julia> @btime overapproximate(B, 1e-2);
  104.749 μs (2631 allocations: 134.19 KiB)

julia> @btime overapproximate(B, 1e-5);
  3.986 ms (87824 allocations: 4.51 MiB)

julia> @btime overapproximate(B, 1e-8);
  221.496 ms (2974378 allocations: 147.37 MiB)

```